+++ src/Entities/PaymentMethodInstrument.php
@@ -11,7 +11,7 @@
 
 namespace Rebilly\Entities;
 
-use DomainException;
+use InvalidArgumentException;
 use Rebilly\Entities\PaymentInstruments\AchInstrument;
 use Rebilly\Entities\PaymentInstruments\AlternativeInstrument;
 use Rebilly\Entities\PaymentInstruments\CashInstrument;
@@ -29,21 +29,6 @@ abstract class PaymentMethodInstrument extends Resource
 
     public const MSG_REQUIRED_METHOD = 'Method is required';
 
-    protected static $supportedCommonMethods = [
-        PaymentMethod::METHOD_ACH,
-        PaymentMethod::METHOD_CASH,
-        PaymentMethod::METHOD_CHECK,
-        PaymentMethod::METHOD_PAYMENT_CARD,
-        PaymentMethod::METHOD_PAYPAL,
-    ];
-
-    protected static $supportedAlternativeMethods = [
-        PaymentMethod::METHOD_IDEAL,
-        PaymentMethod::METHOD_KLARNA,
-        PaymentMethod::METHOD_INTERAC,
-        PaymentMethod::METHOD_ONLINEUEBERWEISEN,
-    ];
-
     /**
      * {@inheritdoc}
      */
@@ -60,7 +45,7 @@ abstract class PaymentMethodInstrument extends Resource
     public static function createFromData(array $data)
     {
         if (!isset($data['method'])) {
-            throw new DomainException(self::MSG_REQUIRED_METHOD);
+            throw new InvalidArgumentException(self::MSG_REQUIRED_METHOD);
         }
 
         switch ($data['method']) {
@@ -75,17 +60,8 @@ abstract class PaymentMethodInstrument extends Resource
             case PaymentMethod::METHOD_PAYPAL:
                 return new PayPalInstrument($data);
             default:
-                if (in_array($data['method'], self::$supportedAlternativeMethods, true)) {
-                    return new AlternativeInstrument($data);
-                }
+                return new AlternativeInstrument($data);
         }
-
-        throw new DomainException(
-            sprintf(
-                self::MSG_UNSUPPORTED_METHOD,
-                implode(',', array_merge(self::$supportedCommonMethods, self::$supportedAlternativeMethods))
-            )
-        );
     }
 
     /**